Make state machine modules auto-loadable
This moves the default state machine modules into `app`, making them
autoloadable. In order to accomplish that, I had to change the
superclass of `Spree::Core::StateMachines` to a
`Spree::Preferences::Configuration`, so that the `class_name_attribute`
method is available. Technically, we're using it for modules here, but
that does not actually matter in this context. We just need it to run
`constantize`.

This also removes some custom "autoloading" code that is not actually
all that necessary. The configuration interface stays the same.
